chore: Separate diff Player controllers
Show the difference between conventional vs MediatR controllers
